requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==2.0.12
dj-database-url==1.2.0
Django==4.2
django-crispy-forms==1.14.0
django-resized==0.3.11
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.6
numpy==1.26.4
openai==0.19.0
openpyxl==3.1.2
pandas==2.2.0
pandas-stubs==2.2.0.240218
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.27.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
types-pytz==2024.1.0.20240203
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
uuid==1.30
whitenoise==6.2.0
